Refactors tests into sensible package hierarchy
