+++ kw_webapp/models.py
@@ -3,6 +3,7 @@ from itertools import chain
 
 from datetime import timedelta
 
+from django.contrib.postgres.fields import JSONField
 from django.db import models
 from django.core.validators import MinValueValidator, MaxValueValidator
 from django.contrib.auth.models import User
